add_deleted_warning_for_vendors_and_runners

+++ Cars/Service/RunnerService.cs
@@ -85,6 +85,11 @@ namespace Cars.Service
                 Runner runner = db.Runners.FirstOrDefault(r => r.RunnerID == RunnerID);
                 if (runner != null)
                 {
+                    var orderlines = db.OrderDetails.Where(or => or.RunnerID == runner.RunnerID);
+                    if (orderlines != null)
+                    {
+                        return -1;
+                    }
                     db.Runners.Remove(runner);
                     db.SaveChanges();
                     return runner.RunnerID;
